+++ project/app/src/main/java/org/owntracks/android/model/messages/MessageClear.kt
@@ -2,8 +2,11 @@ package org.owntracks.android.model.messages
 
 import org.owntracks.android.preferences.Preferences
 import org.owntracks.android.support.Parser
+import java.util.UUID
 
-class MessageClear : MessageBase() {
+class MessageClear : MessageBase(), MessageWithId {
+    override var id: MessageId = ZeroMessageId
+
     override fun toString(): String = "[MessageClear]"
 
     override fun addMqttPreferences(preferences: Preferences) {