+++ project/app/src/main/java/org/owntracks/android/model/messages/MessageCmd.kt
@@ -9,7 +9,7 @@ import org.owntracks.android.preferences.Preferences
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.EXTERNAL_PROPERTY, property = "_type")
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-class MessageCmd : MessageBase() {
+class MessageCmd(private val messageWithId: MessageWithId = MessageWithRandomId()) : MessageBase(), MessageWithId by messageWithId {
     var action: CommandAction? = null
     var waypoints: MessageWaypoints? = null
     var configuration: MessageConfiguration? = null