+++ project/app/src/main/java/org/owntracks/android/model/messages/MessageLwt.kt
@@ -1,6 +1,8 @@
 package org.owntracks.android.model.messages
 
-object MessageLwt : MessageBase() {
+object MessageLwt : MessageBase(), MessageWithId {
     const val TYPE = "lwt"
+    override var id: MessageId = ZeroMessageId
+
     override fun toString(): String = "[MessageLwt]"
 }