+++ project/app/src/main/java/org/owntracks/android/model/messages/MessageTransition.kt
@@ -12,7 +12,7 @@ import org.owntracks.android.preferences.Preferences
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.EXTERNAL_PROPERTY, property = "_type")
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-class MessageTransition : MessageBase() {
+class MessageTransition(private val messageWithId: MessageWithId = MessageWithRandomId()) : MessageBase(), MessageWithId by messageWithId {
 
     @JsonIgnore
     fun getTransition(): Int = when (event) {