+++ project/app/src/main/java/org/owntracks/android/support/IdlingResourceWithData.kt
@@ -1,7 +1,6 @@
 package org.owntracks.android.support
 
 import androidx.test.espresso.IdlingResource
-import java.util.function.Predicate
 import timber.log.Timber
 
 /**
@@ -16,7 +15,8 @@ class IdlingResourceWithData<T>(
     private val comparator: Comparator<in T>
 ) : IdlingResource {
   private var callback: IdlingResource.ResourceCallback? = null
-  private val data = mutableListOf<T>()
+  private val sent = mutableListOf<T>()
+  private val received = mutableListOf<T>()
 
   override fun getName(): String = this.resourceName
 
@@ -24,41 +24,50 @@ class IdlingResourceWithData<T>(
     this.callback = callback
   }
 
-  override fun isIdleNow(): Boolean = data.isEmpty()
+  override fun isIdleNow(): Boolean = sent.isEmpty() && received.isEmpty()
 
   fun add(thing: T) {
-    synchronized(data) {
-      data.add(thing)
-      Timber.v("Waiting for return for $thing")
+    synchronized(sent) {
+      synchronized(received) {
+        Timber.v("Waiting for return for $thing")
+        sent.add(thing)
+        reconcile()
+      }
     }
   }
 
   fun remove(thing: T) {
-    synchronized(data) {
-      val removed = data.removeFirst { comparator.compare(it, thing) == 0 }
-      Timber.v("$name Got return for $thing. Removed=$removed. Remaining=${data.joinToString(",")}")
-      if (data.isEmpty()) {
-        Timber.v("$name Empty. Idling.")
-        callback?.onTransitionToIdle()
+    synchronized(sent) {
+      synchronized(received) {
+        Timber.v("Received return for $thing")
+        received.add(thing)
+        reconcile()
       }
     }
   }
 
-  /**
-   * Removes first item in collection that matches the predicate
-   *
-   * @param filter predicate
-   * @return whether or not an item was removed
-   */
-  private fun <E> MutableCollection<E>.removeFirst(filter: Predicate<in E>): Boolean {
-    val iterator = this.iterator()
-    while (iterator.hasNext()) {
-      val element = iterator.next()
-      if (filter.test(element)) {
-        iterator.remove()
-        return true
+  private fun reconcile() {
+    Timber.v("Contents: sent=${sent.joinToString(",")}, received=${received.joinToString(",")}")
+    sent.intersectByComparator(received.toSet(), comparator).let { sentToRemove ->
+      received.intersectByComparator((sent.toSet()), comparator).let { receivedToRemove ->
+        sent.removeAll(sentToRemove).also {
+          Timber.v("Removed $sentToRemove from sent. Success = $it")
+        }
+        received.removeAll(receivedToRemove).also {
+          Timber.v("Removed $receivedToRemove from received. Success = $it")
+        }
       }
     }
-    return false
+    if (sent.isEmpty() && received.isEmpty()) {
+      Timber.v("$name Empty. Idling.")
+      callback?.onTransitionToIdle()
+    }
+  }
+
+  private fun Collection<T>.intersectByComparator(
+      another: Collection<T>,
+      comparator: Comparator<in T>
+  ): Set<T> {
+    return this.filter { a -> another.any { b -> comparator.compare(a, b) == 0 } }.toSet()
   }
 }