+++ project/app/src/test/java/org/owntracks/android/ui/preferences/load/LoadViewModelTest.kt
@@ -2,11 +2,13 @@ package org.owntracks.android.ui.preferences.load
 
 import android.content.Context
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import com.fasterxml.jackson.databind.ObjectMapper
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -64,14 +66,15 @@ class LoadViewModelTest {
             val preferences = Preferences(preferencesStore, mockIdlingResource)
             val vm = LoadViewModel(preferences, parser, InMemoryWaypointsRepo(), UnconfinedTestDispatcher())
             vm.extractPreferencesFromUri("owntracks:///config?inline=eyJfdHlwZSI6ImNvbmZpZ3VyYXRpb24ifQ")
-            val expectedConfig = """
-            {
-              "_type" : "configuration",
-              "waypoints" : [ ]
-            }
-            """.trimIndent()
+            advanceUntilIdle()
             assertEquals(ImportStatus.SUCCESS, vm.configurationImportStatus.value)
-            assertEquals(expectedConfig, vm.displayedConfiguration.value)
+            val json = ObjectMapper().readTree(vm.displayedConfiguration.value)
+            assertTrue(json.isObject)
+            assertTrue(json.has("_type"))
+            assertEquals("configuration", json.get("_type").asText())
+            assertTrue(json.has("waypoints"))
+            assertTrue(json.get("waypoints").isArray)
+            assertTrue(json.get("waypoints").isEmpty)
         }
 
     @Test
@@ -84,38 +87,30 @@ class LoadViewModelTest {
                 "owntracks:///config?inline=eyJfdHlwZSI6IndheXBvaW50cyIsIndheXBvaW50cyI6W3siX3R5cGUiOiJ3YXlwb2ludCIsImRlc2MiOiJUZXN0IFdheXBvaW50IiwibGF0Ijo1MSwibG9uIjowLCJyYWQiOjQ1MCwidHN0IjoxNTk4NDUxMzcyfV19"
 
             )
-            val expectedConfig = """
-            {
-              "_type" : "waypoints",
-              "waypoints" : [ {
-                "_type" : "waypoint",
-                "desc" : "Test Waypoint",
-                "lat" : 51.0,
-                "lon" : 0.0,
-                "rad" : 450,
-                "tst" : 1598451372
-              } ]
-            }
-            """.trimIndent()
+            val displayedConfig = vm.displayedConfiguration.value
+            val json = ObjectMapper().readTree(displayedConfig)
+            assertTrue(json.isObject)
+            assertTrue(json.has("_type"))
+            assertEquals("waypoints",json.get("_type").asText())
+            assertTrue(json.has("id"))
+            assertTrue(json.get("id").isTextual)
+            assertTrue(json.has("waypoints"))
+            assertTrue(json.get("waypoints").isArray)
+            assertEquals(1, json.get("waypoints").size())
+            assertTrue(json.get("waypoints").get(0).isObject)
+            assertTrue(json.get("waypoints").get(0).has("_type"))
+            assertEquals("waypoint", json.get("waypoints").get(0).get("_type").asText())
+            assertTrue(json.get("waypoints").get(0).has("desc"))
+            assertEquals("Test Waypoint", json.get("waypoints").get(0).get("desc").asText())
+            assertTrue(json.get("waypoints").get(0).has("lat"))
+            assertEquals(51.0, json.get("waypoints").get(0).get("lat").asDouble(),0.000001)
+            assertTrue(json.get("waypoints").get(0).has("lon"))
+            assertEquals(0.0, json.get("waypoints").get(0).get("lon").asDouble(),0.000001)
+            assertTrue(json.get("waypoints").get(0).has("rad"))
+            assertEquals(450, json.get("waypoints").get(0).get("rad").asInt())
+            assertTrue(json.get("waypoints").get(0).has("tst"))
+            assertEquals(1598451372, json.get("waypoints").get(0).get("tst").asLong())
             assertEquals(ImportStatus.SUCCESS, vm.configurationImportStatus.value)
-            assertEquals(expectedConfig, vm.displayedConfiguration.value)
-        }
-
-    @Test
-    fun `Given a configuration ByteArray with a simple configuration, when loading it into the LoadViewModel, then the correct config is displayed`() =
-        runTest {
-            val parser = Parser(null)
-            val preferences = Preferences(preferencesStore, mockIdlingResource)
-            val vm = LoadViewModel(preferences, parser, InMemoryWaypointsRepo(), UnconfinedTestDispatcher())
-            val expectedConfig = """
-            {
-              "_type" : "configuration",
-              "waypoints" : [ ]
-            }
-            """.trimIndent()
-            vm.extractPreferences(expectedConfig.toByteArray())
-            assertEquals(ImportStatus.SUCCESS, vm.configurationImportStatus.value)
-            assertEquals(expectedConfig, vm.displayedConfiguration.value)
         }
 
     @Test