+++ CHANGELOG.md
@@ -31,7 +31,7 @@
 - Contact direction arrow now moves with device to point in the actual direction of the contact
 - Share button added to contact sheet (#1465)
 - Changing the connection details will now clear the contacts and the location message backlog (#1598)
--
+- Messages now include a random `id` (String) field which can be used by any consumer to correlate and distinguish send/return messages
 
 ### Bug fixes